1. 红包列表和奖金列表的金额类型 double -> string

+++ useraccounts-service-provider/src/main/java/com/moseeker/useraccounts/service/impl/biztools/HBBizTool.java
@@ -58,7 +58,7 @@ public class HBBizTool {
                                              Map<Integer, HrHbScratchCardRecord> cardNoMap) {
         RedPacket redPacket = new RedPacket();
         redPacket.setId(hrHbItemsRecord.getId());
-        redPacket.setCandidateName(candidateNameMap.get(hrHbItemsRecord.getTriggerWxuserId()));
+        redPacket.setCandidateName(candidateNameMap.get(hrHbItemsRecord.getTriggerWxuserId().intValue()));
         redPacket.setCardno(cardNoMap.get(hrHbItemsRecord.getId()).getCardno());
         HrHbConfigRecord configRecord = configMap.get(hrHbItemsRecord.getHbConfigId());
         if (configRecord != null) {
